update tong tien, du an id

+++ database/migrations/2023_12_20_012731_create_ton_kho_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(\App\Models\Kho::class, 'kho_id');
             $table->foreignIdFor(\App\Models\SanPham::class, 'san_pham_id');
-            $table->integer('so_luong');
+            $table->double('so_luong',15,2);
             $table->timestamps();
         });
     }
