+++ database/migrations/2024_01_11_110637_add_ton_dau_column_to_nha_cung_cap_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('nha_cung_cap', function (Blueprint $table) {
-            $table->integer('ton_dau')->default(0);
+            $table->double('ton_dau',15,2)->default(0);
         });
     }
