+++ database/migrations/2024_01_18_153136_add_ton_dau_column_to_khach_hang_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('khach_hang', function (Blueprint $table) {
-            $table->integer('ton_dau')->default(0)->after('dia_chi');
+            $table->double('ton_dau',15,2)->default(0)->after('dia_chi');
         });
     }
